# openapi.yml
openapi: 3.0.0
info:
  title: Pipe CLI Web API
  version: 1.0.0
paths:
...
  /api/v1/bff/session-dashboard/{session_id}:
    get:
      summary: Get session dashboard with tree (BFF)
      description: Returns session tree, current session details, and settings in a single request
      tags:
        - v1
        - bff
      x-aggregated-actions:
        - session_tree
        - session/{session_id}
        - settings
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session to retrieve
      responses:
        '200':
          description: Session dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_tree:
                    type: array
                    description: List of all sessions
                    items:
                      type: object
                  current_session:
                    type: object
                    description: Details of the current session
                  settings:
                    type: object
                    description: System settings
                  role_options:
                    type: array
                    description: List of available role options
                    items:
                      $ref: '#/components/schemas/RoleOption'
        '404':
          description: Session not found
        '500':
          description: Internal server error

  # Roles (v1)
...
  /api/v1/session_tree:
    get:
      summary: Get session tree (v1)
      tags:
        - v1
      responses:
        '200':
          description: List of sessions in tree structure
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      type: object
        '500':
          description: Internal server error

  # Session Start (v1)
...
  /api/v1/session/start:
    post:
      summary: Start a new session (singular path used by the Flask app)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
        '422':
          description: Validation error
        '500':
          description: Internal server error
  /api/v1/session/{session_id}:
    get:
      summary: Get a specific session (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session to retrieve
      responses:
        '200':
          description: Session data
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    type: object
        '404':
          description: Session not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a session (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session to delete
      responses:
        '200':
          description: Session deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal server error
  /api/v1/session/{session_id}/turn/{turn_index}:
    delete:
      summary: Delete a specific turn from a session (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session
        - in: path
          name: turn_index
          schema:
            type: integer
          required: true
          description: The index of the turn to delete
      responses:
        '200':
          description: Turn deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Turn index out of range
        '404':
          description: Session not found
        '500':
          description: Internal server error
  /api/v1/session/{session_id}/turns/{turn_index}:
    patch:
      summary: Edit a specific turn in a session (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session
        - in: path
          name: turn_index
          schema:
            type: integer
          required: true
          description: The index of the turn to edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Turn updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: No data provided or turn index out of range
        '403':
          description: Value error
        '404':
          description: Session not found
        '500':
          description: Internal server error
  /api/v1/session/{session_id}/meta:
    patch:
      summary: Edit session metadata (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditSessionMetaRequest'
      responses:
        '200':
          description: Session metadata updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '422':
          description: Validation error
        '404':
          description: Session not found
        '500':
          description: Internal server error
  /api/v1/session/{session_id}/todos:
    patch:
      summary: Edit session todos (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditTodosRequest'
      responses:
        '200':
          description: Session todos updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '422':
          description: Validation error
        '404':
          description: Session not found
        '500':
          description: Internal server error
    delete:
      summary: Delete all todos from a session (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session
      responses:
        '200':
          description: Todos deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Session not found
        '500':
          description: Internal server error
  /api/v1/session/{session_id}/references:
    patch:
      summary: Edit session references (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditReferencesRequest'
      responses:
        '200':
          description: Session references updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '422':
          description: Validation error
        '404':
          description: Session not found
        '500':
          description: Internal server error
  /api/v1/session/{session_id}/references/{reference_index}/persist:
    patch:
      summary: Edit reference persist state (singular path)
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
          required: true
          description: The ID of the session
        - in: path
          name: reference_index
          schema:
            type: integer
          required: true
          description: The index of the reference to edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditReferencePersistRequest'
      responses:
        '200':
          description: Reference persist state updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '422':
          description: Validation error
        '404':
          description: Session not found
        '400':
          description: Reference index out of range
        '500':
          description: Internal server error
...
components:
  schemas:
    RoleOption:
      type: object
      properties:
        label:
          type: string
          description: Display label for the role (e.g., 'python/developer')
        value:
          type: string
          description: Unique identifier for the role (e.g., 'roles/python/developer.md')
      required:
        - label
        - value
    StartSessionRequest:
      type: object
      properties:
        purpose:
          type: string
        background:
          type: string
        instruction:
          type: string
        roles:
          type: array
          items:
            type: string
          nullable: true
        parent:
          type: string
          nullable: true
        references:
          type: array
          items:
            $ref: '#/components/schemas/Reference'
          nullable: true
        artifacts:
          type: array
          items:
            type: string
          nullable: true
        procedure:
          type: string
          nullable: true
        multi_step_reasoning_enabled:
          type: boolean
          default: false
        hyperparameters:
          $ref: '#/components/schemas/Hyperparameters'
          nullable: true
      required:
        - purpose
        - background
        - instruction
    EditSessionMetaRequest:
      type: object
      properties:
        purpose:
          type: string
          nullable: true
        background:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
          nullable: true
        artifacts:
          type: array
          items:
            type: string
          nullable: true
        procedure:
          type: string
          nullable: true
        multi_step_reasoning_enabled:
          type: boolean
          nullable: true
        token_count:
          type: integer
          nullable: true
        hyperparameters:
          $ref: '#/components/schemas/Hyperparameters'
          nullable: true
    EditTodosRequest:
      type: object
      properties:
        todos:
          type: array
          items:
            $ref: '#/components/schemas/TodoItem'
      required:
        - todos
...